+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ### Added
 - #1770: Added missing RDP performance settings
+- #1332: Added option to hide menu strip container
 - #545: Option to minimize to system tray on closing
 - #420: SSH tunneling implemented
 - #283: Support for native PowerShell remoting as new protocol